+++ src/main/java/org/financeiro/dtos/UsuarioDto.java
@@ -1,6 +1,6 @@
 package org.financeiro.dtos;
 
-public record UsuarioDto(int id,
+public record UsuarioDto(long id,
                          String nome,
                          String email,
                          String senha) {
